Moving MPISimultionMode to its own file.
