+++ framework/Simulation.py
@@ -39,7 +39,6 @@ import Files
 import Samplers
 import Optimizers
 import Models
-
 import Metrics
 import Distributions
 import Databases
@@ -117,6 +116,10 @@ class SimulationMode(MessageHandler.MessageUser):
     """
     pass
 
+#Note that this has to be after SimulationMode is defined or the CustomModes
+#don't see SimulationMode when they import Simulation
+import CustomModes
+
 def splitCommand(s):
   """
     Splits the string s into a list that can be used for the command
@@ -188,197 +191,6 @@ def createAndRunQSUB(simulation):
 
 #----------------------------------------------------------------------
 
-class MPISimulationMode(SimulationMode):
-  """
-    MPISimulationMode is a specialized class of SimulationMode.
-    It is aimed to distribute the runs using the MPI protocol
-  """
-  def __init__(self,simulation):
-    """
-      Constructor
-      @ In, simulation, instance, instance of the simulation class
-      @ Out, None
-    """
-    SimulationMode.__init__(self,simulation)
-    self.__simulation = simulation
-    self.messageHandler = simulation.messageHandler
-    #Figure out if we are in PBS
-    self.__inPbs = "PBS_NODEFILE" in os.environ
-    self.__nodefile = False
-    self.__runQsub = False
-    self.__noSplitNode = False #If true, don't split mpi processes across nodes
-    self.__limitNode = False #If true, fiddle with max on Node
-    self.__maxOnNode = None #Used with __noSplitNode and __limitNode to limit number on a node
-    self.__noOverlap = False #Used with __limitNode to prevent multiple batches from being on one node
-    # If (__noSplitNode or __limitNode) and  __maxOnNode is not None,
-    # don't put more than that on on single shared memory node
-    self.printTag = 'MPI SIMULATION MODE'
-
-  def modifySimulation(self):
-    """
-      This method is aimed to modify the Simulation instance in
-      order to distribute the jobs using the MPI protocol
-      @ In, None
-      @ Out, None
-    """
-    if self.__nodefile or self.__inPbs:
-      if not self.__nodefile:
-        #Figure out number of nodes and use for batchsize
-        nodefile = os.environ["PBS_NODEFILE"]
-      else:
-        nodefile = self.__nodefile
-      lines = open(nodefile,"r").readlines()
-      self.__simulation.runInfoDict['Nodes'] = list(lines)
-      numMPI = self.__simulation.runInfoDict['NumMPI']
-      oldBatchsize = self.__simulation.runInfoDict['batchSize']
-      #the batchsize is just the number of nodes of which there is one
-      # per line in the nodefile divided by the numMPI (which is per run)
-      # and the floor and int and max make sure that the numbers are reasonable
-      maxBatchsize = max(int(math.floor(len(lines)/numMPI)),1)
-      if maxBatchsize < oldBatchsize:
-        self.__simulation.runInfoDict['batchSize'] = maxBatchsize
-        self.raiseAWarning("changing batchsize from "+str(oldBatchsize)+" to "+str(maxBatchsize)+" to fit on "+str(len(lines))+" processors")
-      newBatchsize = self.__simulation.runInfoDict['batchSize']
-      if newBatchsize > 1:
-        #need to split node lines so that numMPI nodes are available per run
-        workingDir = self.__simulation.runInfoDict['WorkingDir']
-        if not (self.__noSplitNode or self.__limitNode):
-          for i in range(newBatchsize):
-            nodeFile = open(os.path.join(workingDir,"node_"+str(i)),"w")
-            for line in lines[i*numMPI:
-              (i+1)*numMPI]:
-              nodeFile.write(line)
-            nodeFile.close()
-        else:
-          #self.__noSplitNode == True or self.__limitNode == True
-          nodes = []
-          for line in lines:
-            nodes.append(line.strip())
-
-          nodes.sort()
-
-          currentNode = ""
-          countOnNode = 0
-          nodeUsed = False
-
-          if self.__noSplitNode:
-            groups = []
-          else:
-            groups = [[]]
-
-          for i in range(len(nodes)):
-            node = nodes[i]
-            if node != currentNode:
-              currentNode = node
-              countOnNode = 0
-              nodeUsed = False
-              if self.__noSplitNode:
-                #When switching node, make new group
-                groups.append([])
-            if self.__maxOnNode is None or countOnNode < self.__maxOnNode:
-              countOnNode += 1
-              if len(groups[-1]) >= numMPI:
-                groups.append([])
-                nodeUsed = True
-              if not self.__noOverlap or not nodeUsed:
-                groups[-1].append(node)
-
-          fullGroupCount = 0
-          for group in groups:
-            if len(group) < numMPI:
-              self.raiseAWarning("not using part of node because of partial group: "+str(group))
-            else:
-              nodeFile = open(os.path.join(workingDir,"node_"+str(fullGroupCount)),"w")
-              for node in group:
-                print(node,file=nodeFile)
-              nodeFile.close()
-              fullGroupCount += 1
-          if fullGroupCount == 0:
-            self.raiseAnError(IOError, "Cannot run with given parameters because no nodes have numMPI "+str(numMPI)+" available and NoSplitNode is "+str(self.__noSplitNode)+" and LimitNode is "+str(self.__limitNode))
-          if fullGroupCount != self.__simulation.runInfoDict['batchSize']:
-            self.raiseAWarning("changing batchsize to "+str(fullGroupCount)+" because NoSplitNode is "+str(self.__noSplitNode)+" and LimitNode is "+str(self.__limitNode)+" and some nodes could not be used.")
-            self.__simulation.runInfoDict['batchSize'] = fullGroupCount
-
-        #then give each index a separate file.
-        nodeCommand = self.__simulation.runInfoDict["NodeParameter"]+" %BASE_WORKING_DIR%/node_%INDEX% "
-      else:
-        #If only one batch just use original node file
-        nodeCommand = self.__simulation.runInfoDict["NodeParameter"]+" "+nodefile
-    else:
-      #Not in PBS, so can't look at PBS_NODEFILE and none supplied in input
-      newBatchsize = self.__simulation.runInfoDict['batchSize']
-      numMPI = self.__simulation.runInfoDict['NumMPI']
-      #TODO, we don't have a way to know which machines it can run on
-      # when not in PBS so just distribute it over the local machine:
-      nodeCommand = " "
-
-    #Disable MPI processor affinity, which causes multiple processes
-    # to be forced to the same thread.
-    os.environ["MV2_ENABLE_AFFINITY"] = "0"
-
-    # Create the mpiexec pre command
-    # Note, with defaults the precommand is "mpiexec -f nodeFile -n numMPI"
-    self.__simulation.runInfoDict['precommand'] = self.__simulation.runInfoDict["MPIExec"]+" "+nodeCommand+" -n "+str(numMPI)+" "+self.__simulation.runInfoDict['precommand']
-    if(self.__simulation.runInfoDict['NumThreads'] > 1):
-      #add number of threads to the post command.
-      self.__simulation.runInfoDict['postcommand'] = " --n-threads=%NUM_CPUS% "+self.__simulation.runInfoDict['postcommand']
-    self.raiseAMessage("precommand: "+self.__simulation.runInfoDict['precommand']+", postcommand: "+self.__simulation.runInfoDict['postcommand'])
-
-  def doOverrideRun(self):
-    """
-      If doOverrideRun is true, then use runOverride instead of
-      running the simulation normally.  This method should call
-      simulation.run
-      @ In, None
-      @ Out, doOverrRun, bool, does the override?
-    """
-    # Check if the simulation has been run in PBS mode and if run QSUB
-    # has been requested, in case, construct the proper command
-    doOverrRun = (not self.__inPbs) and self.__runQsub
-    return doOverrRun
-
-  def runOverride(self):
-    """
-      This  method completely overrides the Simulation's run method
-      @ In, None
-      @ Out, None
-    """
-    #Check and see if this is being accidently run
-    assert self.__runQsub and not self.__inPbs
-    createAndRunQSUB(self.__simulation)
-
-  def XMLread(self, xmlNode):
-    """
-      XMLread is called with the mode node, and is used here to
-      get extra parameters needed for the simulation mode MPI.
-      @ In, xmlNode, xml.etree.ElementTree.Element, the xml node that belongs to this class instance
-      @ Out, None
-    """
-    for child in xmlNode:
-      if child.tag == "nodefileenv":
-        self.__nodefile = os.environ[child.text.strip()]
-      elif child.tag == "nodefile":
-        self.__nodefile = child.text.strip()
-      elif child.tag.lower() == "runqsub":
-        self.__runQsub = True
-      elif child.tag.lower() == "nosplitnode":
-        self.__noSplitNode = True
-        self.__maxOnNode = child.attrib.get("maxOnNode",None)
-        if self.__maxOnNode is not None:
-          self.__maxOnNode = int(self.__maxOnNode)
-        if "noOverlap" in child.attrib:
-          self.__noOverlap = True
-      elif child.tag.lower() == "limitnode":
-        self.__limitNode = True
-        self.__maxOnNode = child.attrib.get("maxOnNode",None)
-        if self.__maxOnNode is not None:
-          self.__maxOnNode = int(self.__maxOnNode)
-        else:
-          self.raiseAnError(IOError, "maxOnNode must be specified with LimitNode")
-        if "noOverlap" in child.attrib and child.attrib["noOverlap"].lower() in utils.stringsThatMeanTrue():
-          self.__noOverlap = True
-      else:
-        self.raiseADebug("We should do something with child "+str(child))
 #
 #
 #
@@ -506,7 +318,7 @@ class Simulation(MessageHandler.MessageUser):
 
     #Dictionary of mode handlers for the
     self.__modeHandlerDict           = {}
-    self.__modeHandlerDict['mpi']    = MPISimulationMode
+    self.__modeHandlerDict['mpi']    = CustomModes.MPISimulationMode
 
     #this dictionary contain the static factory that return the instance of one of the allowed entities in the simulation
     #the keywords are the name of the module that contains the specialization of that specific entity
